template/template: Error (was panic) on relative references with nil base
Avoiding the situation discussed in 4d8f93e8 (template: Document
unbased ResolveReference panic, 2017-10-12).  In the Go bug I'd filed,
Ian pointed out that (nil).ResolveReference panics if and only if the
reference scheme is nil, and that IsAbs is only looking at the scheme
[1].  So this commit guards against data that would panic
ResolveReference, instead of catching and handling an error/panic that
ResolveReference raises.

I've also added the base URI to the testcase struct, so we can
exercise different values there.  In this case, I needed to use just
the encoded part to get the relative-reference check to raise an error
at all, because some-algorithm:0123456789abcdef matches the URI syntax
with a 'some-algorithm' scheme and '0123456789abcdef' hier-part [2].
That looks (rightly) absolute, so ResolveReference wasn't looking at
the nil base.  By using just the encoded part, we no longer have the
colon required for an absolute URI.

[1]: https://github.com/golang/go/issues/22229#issuecomment-336231937
[2]: https://tools.ietf.org/html/rfc3986#section-3
[3]: https://tools.ietf.org/html/rfc3986#section-4.2

Signed-off-by: W. Trevor King <wking@tremily.us>