read: Shift registry into a read/ directory
And add a 'get' subcommand to oci-cas.  This gives us room for other
CAS-engine operations, e.g. writers, listers, and deleters.  The
registry is in the 'read' subdirectory, because consumers can use
ref-engine discovery [1,2] and the CAS-engine protocol registry [3]
for reading blobs.  Writing and managing blobs, on the other hand, is
likely to be configured locally, and doesn't need a protocol registry.

[1]: https://github.com/xiekeyang/oci-discovery/blob/0be7eae246ae9a975a76ca209c045043f0793572/well-known-uri-ref-engine-discovery.md
[2]: https://github.com/xiekeyang/oci-discovery/blob/0be7eae246ae9a975a76ca209c045043f0793572/xdg-ref-engine-discovery.md
[3]: https://github.com/xiekeyang/oci-discovery/blob/0be7eae246ae9a975a76ca209c045043f0793572/cas-engine-protocols.md

Signed-off-by: W. Trevor King <wking@tremily.us>