dir: Base Engine in the filesystem root
This allows for template URIs that can access the whole local
filesystem (depending on permissions, etc.).  Limiting makes sense for
configurations loaded from the nework (e.g. via ref-engine discovery),
which is why oci-cas has --file and similar arguments.  But it's less
important for dir.New, because the caller is setting the URI directly,
and can gate-keep as they see fit.

Signed-off-by: W. Trevor King <wking@tremily.us>